JIRA XER-19 Fixed multiframe objects to be correctly displayed.  This uses a fairly optimized back end to store 1 copy with differences in individual images (if any).  If there are no differences, then the storage size per frame server side is 4 bytes (32 bit OS), or 8 bytes (64 bit OS) for an empty pointer.
This code does NOT handle the case of enhanced CT where the object is split into multiple non-contiguous segments, but it does lay the groundwork for that.  To support such an operation, all that would be required would be to extend the existing multi-frame object to one that is an ordered sub-set of the root multi-frame object, and insert one of those objects for each run of frames within the overall series.  This could easily be extended to split MR objects into multiple series as well, as the new objects would be sufficiently distinct to allow them to be in a separate series.