[artf12284]  Multiple changes to allow sharing the layout structure - primarily around using JavaScript as the model instead of Java.
In metadata, made "static" be consistently the values that are used server side to define the script environment.  
shared is the values that need to be copied to the client side and will be available in both server/client models, and model is the dynamic server side model.

Changed almost all of the server side model in the metadata to come from JavaScript instances

Changed the controller actions to be based on filters and the standard metadataservlet so that things are more consistent.

Moved the patient layout definition into the model

Renamed searchController to be FindController
deleted unused patientController

Deleted all the Java implementations of Layout

Made the JS XeroModel be fully-instanatiated in the constructor rather than inline in scripts (so as to be re-useable).

Moved the layout scripts to a layout sub-directory.